+++ NEWS.md
@@ -1,8 +1,38 @@
+# CHANGES IN xfun VERSION 0.36
+
+- Started deprecating `xfun::isFALSE()` in favor of `base::isFALSE()` for R >= 3.5.0 (thanks, @mmaechler, #66); `isFALSE()` will eventually be removed from **xfun** when we do not need to support R < 3.5.0.
+
+# CHANGES IN xfun VERSION 0.35
+
+- Added a new argument `token` to `protect_math()` to optionally include a token around math expressions.
+
+- `base64_uri()` relies less on the **mime** package now. For some common file extensions (e.g., `.jpg`/`.png`), this function knows their MIME types.
+
+- `stringsAsStrings()` has been removed from this package.
+
+# CHANGES IN xfun VERSION 0.34
+
+- Added a new function `alnum_id()` to generate ID strings from a character vector.
+
+- The function `stringsAsStrings()` has been deprecated.
+
+# CHANGES IN xfun VERSION 0.33
+
+- Reverted the change for #68: the characters `-+!_#` are no longer accepted by default in filename extensions, since they are relatively rare and caused a breakage in rstudio/bookdown#1369. If you wish to allow for these characters, you may use the new `extra` argument in `file_ext()` and related functions, e.g., `xfun::file_ext(x, extra = '-+!_#')`.
+
+- The function `stringsAsStrings()` will be deprecated in a future release of **xfun**, because the global option `stringsAsFactors = FALSE` has become the default in base R since 4.0.0.
+
 # CHANGES IN xfun VERSION 0.32
+
+- Added a function `shrink_images()` to shrink images to a maximum width using the **magick** package (thanks, @apreshill, rstudio/blogdown#614).
+
+- Added a function `tinify_dir()` as a wrapper of `tinify()` to compress images under a directory.
+
+- `file_ext()` supports more file extensions now, such as `.c++`, `.FB2K-COMPONENT`, and so on (thanks, @tentacles-from-outer-space, #68).
 
 - Fixed the issue that `xfun::base_pkgs()` could hang R (thanks, @mmaechler, #66).
 
-- Started deprecating `xfun::isFALSE()` in favor of `base::isFALSE()` for R >= 3.5.0 (thanks, @mmaechler, #66); `isFALSE()` will eventually be removed from **xfun** when we do not need to support R < 3.5.0.
+- The `...` argument in `dir_create()` was not passed to `dir.create()`.
 
 # CHANGES IN xfun VERSION 0.31